requirements.txt
langgraph==1.0.5
langchain-google-genai==4.1.3
python-dotenv==1.0.1
reportlab==4.2.5
